refactor: useCreateMutations 커스텀 훅 생성 및 코드 간소화
